Move util functions to dedicated file
